Change: Rename language utils module to lowercase
Comply with out naming scheme. Only modules containing react components
should use camelCase.
