+++ src/web/pages/user-settings/__tests__/helper-functions.test.ts
@@ -5,7 +5,7 @@
 
 import {describe, test, expect} from '@gsa/testing';
 import {getLangNameByCode} from 'web/pages/user-settings/helper-functions';
-import {BROWSER_LANGUAGE} from 'web/utils/Languages';
+import {BROWSER_LANGUAGE} from 'web/utils/languages';
 
 describe('getLangNameByCode', () => {
   test('returns the correct language name for a valid code', () => {
